Rectangle and TestRectangle
1) Rectangle.java, 2) TestRectangle.java


1.Write a Java class called Rectangle that represents a rectangular, two-dimensional region. In
this lab, you learn to overriding a system method. Create a default, empty constructor.

2.Set the double values for the ‘x’ ‘y’ ‘height’ ‘width’ in the Rectangle objects whose bottom-
left corner is specified by the given ‘x’ and ‘y’ coordinates and by the width and height. Your
class will need to use setters and getters with these 4 values. Recall we discussed in class
lectures to put values into the instantiated object using mutators and accessors (setter and
getter) methods.

3.Override the ‘toString()’ base method to produce the following information.
EXAMPLE:
Rectangle [x=2.5, y=13.0, height=14.0, width=4.0] Area is 54.0

4.Write a test program called TestRectangle that creates two objects of the Rectangle class
called R1 and R2. Assign values to the fields of these objects. Print out these Rectangle
objects using System.out.println() from inside the main() method (Hint: this will utilize the
overridden ‘toString()’ in the Rectangle object.

5.Write a method ‘IsOverlap()” that takes two Rectangle objects as parameters and returns
a boolean of true or false. With each Rectangle’s ‘x’ and ‘y’ starting point and each of their
width and height, you need to compute if the two rectangles overlap. Print in the main()
method if R1 and R2 overlap or not.